+++ app/views/businesses/locations/remove.html.erb
