--- app/views/locations/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "locations/address", location: @location
